Slight recognaization
1. Moved prehash into kdf.go
2. segregated very slow group tests with flag to run or not
